Archiving: Restructure directory / namespaces
