+++ src/EventStore.Core/Services/Archive/ArchiveOptions.cs
@@ -1,9 +1,9 @@
 // Copyright (c) Event Store Ltd and/or licensed to Event Store Ltd under one or more agreements.
 // Event Store Ltd licenses this file to you under the Event Store License v2 (see LICENSE.md).
 
-namespace EventStore.Core.Services.Archiver;
+namespace EventStore.Core.Services.Archive;
 
-public class ArchiverOptions {
+public class ArchiveOptions {
 	public bool Enabled { get; init; } = true;
 	public StorageType StorageType { get; init; } = StorageType.None;
 	public FileSystemOptions FileSystem { get; init; } = new();
