+++ src/EventStore.Core/Services/Archive/Archiver/ICommand.cs
@@ -1,7 +1,7 @@
 // Copyright (c) Event Store Ltd and/or licensed to Event Store Ltd under one or more agreements.
 // Event Store Ltd licenses this file to you under the Event Store License v2 (see LICENSE.md).
 
-namespace EventStore.Core.Services.Archiver;
+namespace EventStore.Core.Services.Archive.Archiver;
 
 public interface ICommand;
 public record Command : ICommand;
