+++ src/EventStore.Core/Services/Archive/Storage/Exceptions/ChunkDeletedException.cs
@@ -1,8 +1,8 @@
 // Copyright (c) Event Store Ltd and/or licensed to Event Store Ltd under one or more agreements.
 // Event Store Ltd licenses this file to you under the Event Store License v2 (see LICENSE.md).
 
-namespace EventStore.Core.Services.Archiver.Storage;
+using System;
 
-public interface IArchiveStorageFactory {
-	IArchiveStorage Create();
-}
+namespace EventStore.Core.Services.Archive.Storage.Exceptions;
+
+public class ChunkDeletedException : Exception;
