+++ src/EventStore.Core/Services/Archive/Storage/IArchiveStorageFactory.cs
@@ -1,8 +1,8 @@
 // Copyright (c) Event Store Ltd and/or licensed to Event Store Ltd under one or more agreements.
 // Event Store Ltd licenses this file to you under the Event Store License v2 (see LICENSE.md).
 
-using System;
+namespace EventStore.Core.Services.Archive.Storage;
 
-namespace EventStore.Core.Services.Archiver.Storage.Exceptions;
-
-public class ChunkDeletedException : Exception;
+public interface IArchiveStorageFactory {
+	IArchiveStorage Create();
+}
